Moved Configuration class to seperate file
